% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellColors.R
\name{cellColors}
\alias{cellColors}
\title{Return default colors for the cell types}
\usage{
cellColors()
}
\value{
A named vector. Due to existance of different naming schemes, 
same cell type can be referred to multiple times.
}
\description{
Return default colors for the cell types
}

