% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeChips.R
\name{mergeChips}
\alias{mergeChips}
\title{Merge two affymetrix chips}
\usage{
mergeChips(affy1, affy2, allowIntersect = FALSE)
}
\arguments{
\item{affy1}{An \link[Biobase]{AffyBatch} object}

\item{affy2}{Another \link[Biobase]{AffyBatch} object, sharing probes with aff1}

\item{allowIntersect}{If TRUE, it will not return an error when one of the chips 
do not include all probesets from the other.}
}
\description{
Takes two affy \link[Biobase]{AffyBatch} objects and converts them 
into an \link[Biobase]{ExpressionSet} object. It subsets the probes with the same
identifiers and discards any probes that are not shared between the two chips.
Intended use is merging data from chips and their later versions.
}

