% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDesignMergeCel.R
\name{readDesignMergeCel}
\alias{readDesignMergeCel}
\title{Read the cell type study metadata file and create the RMA normalized expression matrix.}
\usage{

  readDesignMergeCel(desFile = "data-raw/Mouse_Cell_Type_Data/cellTypeStudies.tsv",
  gsm = "samples", normalize,
  celRegex = "(GSM.*?(?=,|$))|(PC\\\\d....)|(Y[+].*?((?=(,))|\\\\d+))|((?<=:)|(?<=[,]))A((9)|(10))_[0-9]{1,}_Chee_S1_M430A|(v2_(?![G,H,r]).*?((?=(,))|($)))|(SSC.*?((?=(,))|($)))|(MCx.*?((?=(,))|($)))|(Cbx.*?((?=(,))|($)))",
  celDir = "data-raw/cel/", expFile, desOut,
  gemmaDir = "data-raw/GemmaAnnots")
}
\arguments{
\item{desFile}{Address of the study metadata file}

\item{gsm}{column name that has the GSM identifiers of the samples. The identifiers must
be sperated by commas}

\item{normalize}{The column name that marks which studies should be included.}

\item{celRegex}{Regular expression that picks the sample names from the gsm column. 
Default option is for the cellTypeStudies.tsv. Any modifications to the file should not
require changes to the default as long as they are samples from GEO with GSM identifiers.
If files that are named differently are added, modify the regex to recognize them.}

\item{expFile}{name of the output file for gene expression. The file will be a csv}

\item{desOut}{name of the output file for sample metadata. The file will be a tsv}

\item{gemmaDir}{The directory that houses the gemma annotations that are created by \link[ogbox]{getGemmaAnnot}}
}
\description{
This function reads the study design file 'data-raw/Mouse_Cell_Type_Data/cellTypeStudies.tsv' 
included in the package and cel files pointed in the file to create the RMA normalized
 expression matrix. The products of this function are already loaded with the package. 
Use only for replication or upon editing cellTypeStudies.tsv for personal use.
}
