% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{regionHierarchy}
\alias{regionHierarchy}
\title{Region hierarchy}
\format{An object of class \code{list} of length 1.}
\usage{
regionHierarchy
}
\description{
This list holds region tree used in the study and neuroexpresso.org
}
\keyword{datasets}
